+++ src/alphadb/utils/common.py
@@ -13,4 +13,9 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-CONFIG_TABLE_NAME = "adb_conf"
+def convert_version_number(version_number: str):
+    
+    try:
+        return int(version_number.replace(".", ""))
+    except:
+        raise ValueError(f'Version number "{version_number}" is not convertable to an integer.')