+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - Better error handling.
- - Altertable option to update method.
+ - `Altertable` option to update method.
  - Version source validation
 
 ### Changed